test(input): Move test to directive tests
